moved CM provisioning projects to CM folder
