+++ sdk/cloudmachine/Azure.CloudMachine.All/src/Azure.CloudMachine.All.csproj
@@ -10,7 +10,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\..\..\provisioning\Azure.Provisioning.CloudMachine\src\Azure.Provisioning.CloudMachine.csproj" />
+    <ProjectReference Include="..\..\Azure.Provisioning.CloudMachine\src\Azure.Provisioning.CloudMachine.csproj" />
     <ProjectReference Include="..\..\Azure.CloudMachine.Web\src\Azure.CloudMachine.Web.csproj" />
     <ProjectReference Include="..\..\Azure.CloudMachine\src\Azure.CloudMachine.csproj" />
   </ItemGroup>
