+++ sdk/cloudmachine/Azure.CloudMachine/src/Core/LoggingPolicy.cs
