+++ sdk/cloudmachine/Azure.CloudMachine/src/Core/RestCLient/RestCallFailedException.cs
@@ -6,8 +6,16 @@ using System.ClientModel.Primitives;
 
 namespace Azure
 {
+    /// <summary>
+    /// Exception thrown when a REST call fails.
+    /// </summary>
     public class RestCallFailedException : Exception
     {
+        /// <summary>
+        /// Initializes a new instance of the <see cref="RestCallFailedException"/> class.
+        /// </summary>
+        /// <param name="message"></param>
+        /// <param name="response"></param>
         public RestCallFailedException(string message, PipelineResponse response)
             : base(message, new System.ClientModel.ClientResultException(response)) { }
     }
