+++ sdk/cloudmachine/Azure.Provisioning.CloudMachine/src/Azure.Provisioning.CloudMachine.csproj
@@ -5,12 +5,13 @@
     <Version>1.0.0-beta.1</Version>
     <TargetFrameworks>$(RequiredTargetFrameworks)</TargetFrameworks>
     <LangVersion>12</LangVersion>
-
+    <Nullable>enable</Nullable>
     <!-- Disable warning CS1591: Missing XML comment for publicly visible type or member -->
     <NoWarn>CS1591;AZC0007</NoWarn>
   </PropertyGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\..\Azure.CloudMachine\src\Azure.CloudMachine.csproj" />
     <PackageReference Include="Azure.Identity" />
     <PackageReference Include="Azure.Provisioning" />
     <PackageReference Include="Azure.Provisioning.KeyVault" />
