+++ sdk/cloudmachine/Azure.Provisioning.CloudMachine/src/AzureSdkExtensions/AppServiceFeature.cs
