+++ sdk/cloudmachine/Azure.Provisioning.CloudMachine/src/AzureSdkExtensions/KeyVaultFeature.cs
