+++ sdk/cloudmachine/Azure.Provisioning.CloudMachine/src/AzureSdkExtensions/OpenAIFeature.cs
