+++ sdk/cloudmachine/Azure.Provisioning.CloudMachine/src/AzureSdkExtensions/OpenAIModelFeature.cs
