+++ sdk/cloudmachine/Azure.Provisioning.CloudMachine/src/CDKLevel3/Azd.cs
@@ -1,0 +1,82 @@
+﻿// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT License.
+
+using System.IO;
+using Azure.Provisioning;
+using Azure.Provisioning.Primitives;
+using Azure.Provisioning.Resources;
+using Azure.Provisioning.Expressions;
+
+namespace Azure.CloudMachine;
+
+/// <summary>
+/// Azure Developer CLI helpers.
+/// </summary>
+#pragma warning disable AZC0012 // Avoid single word type names
+public static class Azd
+#pragma warning restore AZC0012 // Avoid single word type names
+{
+    private const string MainBicepName = "main";
+    private const string ResourceGroupVersion = "2024-03-01";
+
+    public static void Init(CloudMachineInfrastructure infra, string? infraDirectory = default)
+    {
+        if (infraDirectory == default) infraDirectory = Path.Combine(".", "infra");
+
+        Directory.CreateDirectory(infraDirectory);
+
+        infra.Build().Save(infraDirectory);
+        var cmid = AppConfigHelpers.ReadOrCreateCmid();
+
+        // main.bicep
+        var location = new ProvisioningParameter("location", typeof(string));
+        var principalId = new ProvisioningParameter("principalId", typeof(string));
+
+        ResourceGroup rg = new(nameof(rg), ResourceGroupVersion)
+        {
+            Name = cmid,
+            Location = location
+        };
+
+        Infrastructure mainBicep = new("main")
+        {
+            TargetScope = DeploymentScope.Subscription
+        };
+        ModuleImport import = new("cm", $"cm.bicep")
+        {
+            Name = "cm",
+            Scope = new IdentifierExpression(rg.BicepIdentifier)
+        };
+        import.Parameters.Add(nameof(location), location);
+        import.Parameters.Add(nameof(principalId), principalId);
+
+        mainBicep.Add(rg);
+        mainBicep.Add(import);
+        mainBicep.Add(location);
+        mainBicep.Add(principalId);
+        mainBicep.Build().Save(infraDirectory);
+
+        WriteMainParametersFile(infraDirectory);
+    }
+    private static void WriteMainParametersFile(string infraDirectory)
+    {
+        File.WriteAllText(Path.Combine(infraDirectory, $"{MainBicepName}.parameters.json"),
+        """
+        {
+          "$schema": "https://schema.management.azure.com/schemas/2019-04-01/deploymentParameters.json#",
+          "contentVersion": "1.0.0.0",
+          "parameters": {
+            "environmentName": {
+              "value": "${AZURE_ENV_NAME}"
+            },
+            "location" : {
+              "value" : "${AZURE_LOCATION}"
+            },
+            "principalId": {
+              "value": "${AZURE_PRINCIPAL_ID}"
+            }
+          }
+        }
+        """);
+    }
+}
