+++ sdk/cloudmachine/Azure.Provisioning.CloudMachine/src/CDKLevel3/CloudMachineFeature.cs
