+++ sdk/cloudmachine/Azure.Provisioning.CloudMachine/src/CDKLevel3/FeatureCollection.cs
