+++ sdk/cloudmachine/Azure.Provisioning.CloudMachine/src/Properties/AssemblyInfo.cs
