+++ sdk/cloudmachine/Azure.Provisioning.CloudMachine/src/Tsp/TypeSpecWriter.cs
