+++ sdk/cloudmachine/Azure.Provisioning.CloudMachine/tests/CloudMachineTests.cs
