+++ sdk/cloudmachine/Azure.Provisioning.CloudMachine/tests/appsettings.json
